requirements.txt
humanize==0.5.1
Keras==2.1.5
lxml==4.2.5
msgpack==0.5.6
mock==2.0.0
numpy==1.15.4
pandas==0.23.4
py2neo==4.1.3
pycorenlp==0.3.0
pytest==4.0.2
python-dateutil==2.7.5
PyYAML==3.13
ray==0.6.0
spacy==2.0.18
tableschema==1.3.0
tensorflow==1.6.0
tika==1.19